requirements.txt
fastapi
httpx
uvicorn
pydantic
sqlalchemy
alembic
pydantic-settings
psycopg2-binary
python-decouple
python-multipart
pyjwt
passlib[bcrypt]
pydantic[email]
uvicorn[standard]
websockets